Feat: 시간의 함수로 수식계산해서 표시하도록 코드 수정
 - 테스트돌리는게 잘 안돼서 일단 저장 차원에서 올립니다


Former-commit-id: 426b2208777dac639645f4a9196886cf53dbd829 [formerly dbca1fcb66fb62569dc6922f7611ac422738860d]
Former-commit-id: 34dce46ffaedcc72c08184b472dd82859c56d954

+++ Assets/Scripts/InGameScene/Utils/Equation.cs
@@ -13,14 +13,14 @@ public class Equation
     /// <param name="order"></param>
     /// <param name="elapsedTime"></param>
     /// <returns></returns>
-    public static List<UpArrowNotation> GetTimeCoeffs(int order, int elapsedTime)
+    public static List<UpArrowNotation> GetTimeCoeffs(int order, double elapsedTime)
     {
         List<UpArrowNotation> resultList = new List<UpArrowNotation>();
         resultList.Add(new UpArrowNotation(1));
 
-        for (int i=0; i<order; i++)
+        for (int i = 0; i < order; i++)
         {
-            resultList.Add((resultList[resultList.Count-1] * elapsedTime) / resultList.Count);
+            resultList.Add((resultList[resultList.Count - 1] * elapsedTime) / resultList.Count);
         }
 
         return resultList;
@@ -35,7 +35,7 @@ public class Equation
     {
         UpArrowNotation result = new UpArrowNotation();
 
-        for (int i=0; i<brainEqCoeffs.Count; i++)
+        for (int i = 0; i < brainEqCoeffs.Count; i++)
         {
             result += brainEqCoeffs[i] * timeCoeffs[i];
         }
@@ -43,8 +43,8 @@ public class Equation
         return result;
     }
 
-    public static UpArrowNotation GetCurrentIntellect(List<UpArrowNotation> brainEqCoeffs, int elapsedTime)
+    public static UpArrowNotation GetCurrentIntellect(List<UpArrowNotation> brainEqCoeffs, double elapsedTime)
     {
-        return MultiplyCoeffs(brainEqCoeffs, GetTimeCoeffs(brainEqCoeffs.Count-1, elapsedTime));
+        return MultiplyCoeffs(brainEqCoeffs, GetTimeCoeffs(brainEqCoeffs.Count - 1, elapsedTime));
     }
 }
